Ball radius created
The ball class now has a radius variable that it has to adhere to.
Note!: The radius gets implemented to the ball a bit late in the rendering process, so I feel like it might bite me in the ass later. But considering the main focus of this assignment is a collision engine, I do not feel like I should - nor that I need to - tweak right now.