Pseudo-bounces
There is a fault somewhere in the formula (I assume it is either the set-velocity formula or something to do with d). But Using the Ball-Wall bounce, the balls bounce as if they hit the wall, while using the normal found in the formula to figure out what direction to bounce.

The balls bounce as if they hit a wall, so in a few instances the balls bounce abnormally.