Entity Component System v3.0
SURELY, this one will work!
Created an Entity builder to create the connection between an entity and a transform component.
Component system is completely revamped.
Components are now placed in the entity.
Creating an entity and (supposedly with) a component attached to it DOESN'T CRASH THE PROGRAM (likely due to pointer errors being fixed)!